requirements.txt
python-dotenv
streamlit
langchain
groq
pypdf2